+++ pnet-data-api-java/src/main/java/pnet/data/api/activity/ActivityDataFind.java
@@ -2,7 +2,7 @@ package pnet.data.api.activity;
 
 import java.util.List;
 
-import pnet.data.api.util.AbstractFind;
+import pnet.data.api.util.AbstractScrollableFind;
 import pnet.data.api.util.FindFunction;
 import pnet.data.api.util.IncludeInactive;
 import pnet.data.api.util.Pair;
@@ -21,7 +21,7 @@ import pnet.data.api.util.Scrollable;
  *
  * @author ham
  */
-public class ActivityDataFind extends AbstractFind<ActivityItemDTO, ActivityDataFind>
+public class ActivityDataFind extends AbstractScrollableFind<ActivityItemDTO, ActivityDataFind>
     implements RestrictMatchcode<ActivityDataFind>, RestrictTenant<ActivityDataFind>, RestrictBrand<ActivityDataFind>,
     RestrictCompanyType<ActivityDataFind>, RestrictContractType<ActivityDataFind>, RestrictVisibility<ActivityDataFind>,
     RestrictApprovalNeeded<ActivityDataFind>, RestrictUpdatedAfter<ActivityDataFind>, IncludeInactive<ActivityDataFind>,