+++ pnet-data-api-java/src/main/java/pnet/data/api/application/ApplicationDataFind.java
@@ -2,7 +2,7 @@ package pnet.data.api.application;
 
 import java.util.List;
 
-import pnet.data.api.util.AbstractFind;
+import pnet.data.api.util.AbstractScrollableFind;
 import pnet.data.api.util.FindFunction;
 import pnet.data.api.util.IncludeInactive;
 import pnet.data.api.util.Pair;
@@ -13,7 +13,7 @@ import pnet.data.api.util.Scrollable;
 /**
  * Find for applications
  */
-public class ApplicationDataFind extends AbstractFind<ApplicationItemDTO, ApplicationDataFind>
+public class ApplicationDataFind extends AbstractScrollableFind<ApplicationItemDTO, ApplicationDataFind>
     implements RestrictMatchcode<ApplicationDataFind>, RestrictUpdatedAfter<ApplicationDataFind>,
     IncludeInactive<ApplicationDataFind>, Scrollable<ApplicationDataFind>
 {